improve typing on fastio
